[eudsl-llvmpy] ILP: address PR #671 review comments
- Fold mir_ilp_model.py into mir_ilp_base.py (the pure helpers now live beside
  the base they serve; still ortools-free at import).
- make_solver: run the CP-SAT portfolio across (ncpu - 2) workers instead of a
  single worker.
- RAILPPacking._solve: add a formal statement of the 2D-packing ILP (variables,
  rectangles, no-overlap constraint, spill indicator, objective).
- Move the comparison runner from scripts/ to tests/mir/.
- Drop the comparison-runner blurb from the README and two now-obvious inline
  comments from the CI workflow.
