+++ projects/eudsl-llvmpy/src/llvm/mir_ilp_base.py
@@ -14,12 +14,118 @@ assignment is a hard error, so a model bug surfaces instead of being silently
 repaired (which would make the ILP-vs-greedy comparison meaningless). Only
 reload vregs the ILP never saw (minted by ``self.spill`` after the solve) are
 colored first-free.
+
+This module also holds the pure, binding-free helpers (weight scaling, live-
+range interference, time-axis compaction, candidate filtering, single-class
+detection) shared by the concrete ``_solve`` models. They take plain Python
+data so they are unit-testable without a MachineFunction, and the CP-SAT module
+is imported lazily via ``_require_ortools`` so the package imports without the
+optional ``ortools`` dependency.
 """
 
+import os
 from dataclasses import dataclass, field
 
 from . import mir
-from .mir_ilp_model import scale_weight
+
+# ---- pure helpers (weights, interference, time axis, register classes) ----
+
+# Spill weights are floats; CP-SAT objectives are integer. Scale by this factor
+# so the weighted-spill objective dominates the unit coalescing-hint bonus (a
+# hint can only ever break ties, never force or prevent a spill).
+_WEIGHT_SCALE = 1000
+
+# LLVM marks must-not-spill intervals with an infinite (HUGE_VALF) spill weight.
+# CP-SAT needs a finite integer coefficient, so clamp to a value large enough to
+# dominate any realistic sum of ordinary weights yet safely within int range.
+_MAX_WEIGHT = 1_000_000_000
+
+# Reward (in scaled objective units) for assigning a vreg to its copy hint.
+# Strictly less than the smallest scaled spill weight so it never buys a spill.
+HINT_BONUS = 1
+
+
+def _require_ortools():
+    try:
+        from ortools.sat.python import cp_model
+    except ImportError as e:  # pragma: no cover - exercised only without the extra
+        raise ImportError(
+            "ortools is required for the ILP register allocators; install "
+            "eudsl-llvmpy[ilp] or `pip install ortools>=9.0`"
+        ) from e
+    return cp_model
+
+
+def scale_weight(weight):
+    """Scale a float spill weight to a positive integer objective coefficient.
+
+    LLVM's must-not-spill marker (infinite / HUGE_VALF weight) maps to
+    _MAX_WEIGHT. Every finite weight maps strictly below it (capped at
+    _MAX_WEIGHT - 1), so an unspillable interval always dominates any finite
+    spill cost while distinct finite weights stay distinguishable up to the cap.
+    """
+    if weight == float("inf"):
+        return _MAX_WEIGHT
+    return max(1, min(_MAX_WEIGHT - 1, round(weight * _WEIGHT_SCALE)))
+
+
+def _segments_overlap(segs_a, segs_b):
+    """True if any half-open [start, end) segment of A overlaps one of B."""
+    for s1, e1 in segs_a:
+        for s2, e2 in segs_b:
+            if s1 < e2 and s2 < e1:
+                return True
+    return False
+
+
+def build_interference(intervals):
+    """Pairwise interference edges from live-range overlap.
+
+    `intervals` maps vreg id -> list of (start, end) half-open integer segments.
+    Returns a set of ``frozenset({u, v})`` for every pair whose ranges overlap.
+    """
+    vregs = sorted(intervals)
+    edges = set()
+    for i, u in enumerate(vregs):
+        for v in vregs[i + 1 :]:
+            if _segments_overlap(intervals[u], intervals[v]):
+                edges.add(frozenset((u, v)))
+    return edges
+
+
+def compact_time_axis(intervals):
+    """Map the sorted set of all segment endpoints to contiguous ints.
+
+    Returns ``(mapping, n_points)`` where `mapping` sends each original endpoint
+    to its index in the sorted order. Used to give the packing model a dense
+    time axis instead of raw (possibly large, sparse) slot distances.
+    """
+    points = sorted({p for segs in intervals.values() for seg in segs for p in seg})
+    mapping = {p: i for i, p in enumerate(points)}
+    return mapping, len(points)
+
+
+def candidate_pregs(order, forbidden):
+    """Allocation-order physregs minus the matrix-forbidden ones."""
+    return [p for p in order if p not in forbidden]
+
+
+def single_class_k(reg_class_id, num_regs):
+    """If every vreg shares one register class, return its allocatable count k;
+    else None. The packing and decomposition models are scoped to single-class
+    functions.
+
+    Gating on class *identity* (not merely on the allocatable count) is what
+    makes the flat register axis sound: two different classes can report the
+    same count yet alias. On AArch64, GPR32 and GPR64 both have 30 allocatable
+    registers, but W0/X0 share register units -- treating them as independent
+    axis indices would silently ignore that aliasing. `reg_class_id` maps each
+    vreg to its TargetRegisterClass id; `num_regs` maps it to the class's
+    allocatable-register count (identical across vregs when the class is).
+    """
+    if len(set(reg_class_id.values())) != 1:
+        return None
+    return next(iter(num_regs.values()))
 
 
 @dataclass
@@ -80,16 +186,19 @@ def stats_from_solver(cp, solver, status, wall):
 
 
 def make_solver(cp, time_limit_s):
-    """A CP-SAT solver configured for reproducible, deterministic search.
+    """A CP-SAT solver: fixed seed, parallel portfolio across ``ncpu - 2``.
 
-    A fixed seed and a single worker make allocation reproducible across runs
-    (the study compares solutions, and reproducible compiler output matters);
-    the problems here are small enough that single-threaded search is fine.
+    The fixed seed keeps the search order reproducible; the solver runs its
+    strategy portfolio over ``ncpu - 2`` workers (at least one) to solve faster.
+    For the small study problems here the search reaches a proven optimum well
+    within the time limit, so the objective/gap are stable; note that with more
+    than one worker the *particular* optimum chosen among ties (and any
+    incumbent returned on timeout) may vary run to run.
     """
     solver = cp.CpSolver()
     solver.parameters.max_time_in_seconds = time_limit_s
     solver.parameters.random_seed = 0
-    solver.parameters.num_workers = 1
+    solver.parameters.num_workers = max(1, (os.cpu_count() or 1) - 2)
     return solver
